mqtt_client/requirements.txt
paho-mqtt==1.6.1
requests==2.26.0
channels==4.0.0
channels-redis==4.0.0
redis==4.4.0
asgiref==3.6.0
pytz==2023.3
xmltodict==0.13.0
pandas==1.5.2
python-dotenv == 0.19.1
